+++ system/src/Grav/Framework/Flex/Storage/SimpleStorage.php
@@ -13,6 +13,7 @@ namespace Grav\Framework\Flex\Storage;
 
 use Grav\Common\Data\Data;
 use Grav\Common\Filesystem\Folder;
+use Grav\Common\Utils;
 use Grav\Framework\Filesystem\Filesystem;
 use InvalidArgumentException;
 use LogicException;
@@ -53,9 +54,9 @@ class SimpleStorage extends AbstractFilesystemStorage
         $filesystem = Filesystem::getInstance(true);
 
         $extension = $this->dataFormatter->getDefaultFileExtension();
-        $pattern = basename($options['folder']);
+        $pattern = Utils::basename($options['folder']);
 
-        $this->dataPattern = basename($pattern, $extension) . $extension;
+        $this->dataPattern = Utils::basename($pattern, $extension) . $extension;
         $this->dataFolder = $filesystem->dirname($options['folder']);
         $this->keyField = $options['key'] ?? 'storage_key';
         $this->keyLen = (int)($options['key_len'] ?? 32);
@@ -432,7 +433,7 @@ class SimpleStorage extends AbstractFilesystemStorage
      */
     protected function getKeyFromPath(string $path): string
     {
-        return basename($path);
+        return Utils::basename($path);
     }
 
     /**